test(source-patch): prove ON OFF product cycles
